F haki

+++ youconnect/app/Http/Controllers/PostController.php
@@ -17,9 +17,9 @@ class PostController extends Controller
         return view('profiles.Myposts', compact('posts', 'user'));
     }
 
-    public function view(Post $post)
+    public function updatePost(Post $post)
     {
-        return view('main.view', compact('post','users'));
+        return view('main.update', compact('post'));
     }
 
     public function store(Request $request)
@@ -40,10 +40,7 @@ class PostController extends Controller
         return redirect(route('index'))->withSuccess('Successfully made post!');
     }
 
-    public function edit(Post $post)
-    {
-        return view('main.post.edit', ['post' => $post]);
-    }
+    
     public function update(Request $request, Post $post)
     {
         $data = $request->validate([
@@ -60,14 +57,19 @@ class PostController extends Controller
         }
 
         $post->update($data);
-        return redirect(route('main.posts'))->withSuccess('Successfully updated post!');
+        return redirect(route('index'))->withSuccess('Successfully updated post!');
     }
-
+            
     public function destroy(Post $post)
-    {
+{
+    if ($post->user->id === Auth::user()->id) {
         $post->delete();
-        return redirect(route('main.posts'))->withSuccess('Deleted successfully!');
+        return redirect(route('index'))->withSuccess('Post deleted successfully!');
+    } else {
+        return redirect(route('index'))->withError('You are not authorized to delete this post!');
     }
+}
+
 
     
 }
